+++ tlRender/lib/tlVk/Shader.h
@@ -53,11 +53,12 @@ namespace tl
             void _init(const std::string& vertexSource,
                        const std::string& fragmentSource,
                        const std::string& name);
-            void _init(const uint8_t* vertexBytes, const uint32_t vertexLength,
+            void _init(const uint32_t* vertexBytes, const uint32_t vertexLength,
                        const std::string& fragmentSource,
                        const std::string& name);
-            void _init(const uint8_t* vertexBytes, const uint32_t vertexLength,
-                       const uint8_t* fragmentBytes, const uint32_t fragmentLength,
+            void _init(const uint32_t* vertexBytes, const uint32_t vertexLength,
+                       const uint32_t* fragmentBytes,
+                       const uint32_t fragmentLength,
                        const std::string& name);
 
 
@@ -71,7 +72,7 @@ namespace tl
             //! Create a new shader.
             static std::shared_ptr<Shader> create(
                 Fl_Vk_Context& ctx,
-                const uint8_t* vertexBytes,
+                const uint32_t* vertexBytes,
                 const uint32_t vertexLength,
                 const std::string& fragmentSource,
                 const std::string& name = "");
@@ -79,9 +80,9 @@ namespace tl
             //! Create a new shader.
             static std::shared_ptr<Shader> create(
                 Fl_Vk_Context& ctx,
-                const uint8_t* vertexBytes,
+                const uint32_t* vertexBytes,
                 const uint32_t vertexLength,
-                const uint8_t* fragmentBytes,
+                const uint32_t* fragmentBytes,
                 const uint32_t fragmentLength,
                 const std::string& name = "");
